Added the ability to like posts

+++ post-service/src/main/java/ru/andreyszdlv/postservice/service/LikeService.java
@@ -4,8 +4,11 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.andreyszdlv.postservice.api.userservice.UserServiceFeignClient;
 import ru.andreyszdlv.postservice.exception.AlreadyLikedException;
+import ru.andreyszdlv.postservice.exception.NoLikedPostThisUserException;
+import ru.andreyszdlv.postservice.exception.NoSuchPostException;
 import ru.andreyszdlv.postservice.model.Like;
 import ru.andreyszdlv.postservice.repository.LikeRepo;
+import ru.andreyszdlv.postservice.repository.PostRepo;
 
 @Service
 @AllArgsConstructor
@@ -13,9 +16,14 @@ public class LikeService {
 
     private final LikeRepo likeRepository;
 
+    private final PostRepo postRepository;
+
     private final UserServiceFeignClient userServiceFeignClient;
 
     public void createLike(long postId){
+        if(!postRepository.existsById(postId)) {
+            throw new NoSuchPostException("errors.404.post_not_found");
+        }
         long userId = userServiceFeignClient.getUserIdByUserEmail().getBody();
         if(likeRepository.existsByPostIdAndUserId(postId, userId)){
             throw new AlreadyLikedException("errors.409.already_liked_post");
@@ -27,6 +35,13 @@ public class LikeService {
     }
 
     public void deleteLike(long postId) {
+        if(!postRepository.existsById(postId)) {
+            throw new NoSuchPostException("errors.404.post_not_found");
+        }
+        long userId = userServiceFeignClient.getUserIdByUserEmail().getBody();
+        if(!likeRepository.existsByPostIdAndUserId(postId, userId)){
+            throw new NoLikedPostThisUserException("errors.404.no_liked_post");
+        }
         likeRepository.deleteByPostIdAndUserId(postId, userServiceFeignClient.getUserIdByUserEmail().getBody());
     }
 }
